feat(cart): add calculatedPrice for each dto for product, customRule, option, and trait

+++ src/main/java/com/whattheburger/backend/service/dto/ProductDto.java
@@ -61,7 +61,7 @@ public class ProductDto {
         private Long optionTraitId;
         private String name;
         private Integer defaultSelection; //productOptionTrait
-        private Double extraPrice; //productOptionTrait
+        private BigDecimal extraPrice; //productOptionTrait
         private Double extraCalories; //productOptionTrait
     }
 }
